requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
praw==7.7.1
yfinance==0.2.43
pytrends==4.9.2
requests==2.32.3
python-multipart==0.0.9
aiofiles==24.1.0
apscheduler==3.10.4
pandas==2.2.3
numpy==2.1.1
python-dotenv==1.0.1
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
